move some utils from engine-api-client to @lodestar/utils
