+++ packages/beacon-node/src/chain/produceBlock/produceBlockBody.ts
@@ -23,9 +23,9 @@ import {
 } from "@lodestar/state-transition";
 import {IChainForkConfig} from "@lodestar/config";
 import {ForkName} from "@lodestar/params";
-import {toHex, sleep} from "@lodestar/utils";
+import {toHex, sleep, numToQuantity} from "@lodestar/utils";
 
-import {IExecutionBuilder, IExecutionEngine, numToQuantity, PayloadId} from "@lodestar/engine-api-client";
+import {IExecutionBuilder, IExecutionEngine, PayloadId} from "@lodestar/engine-api-client";
 import type {BeaconChain} from "../chain.js";
 import {ZERO_HASH, ZERO_HASH_HEX} from "../../constants/index.js";
 import {IEth1ForBlockProduction} from "../../eth1/index.js";
